# backend/docs/swagger.yaml
definitions:
...
  models.ErrorResponse:
    properties:
      '{route} error':
        example: '{method}: {crud}: error message'
        type: string
    type: object
  models.JWTModel:
    properties:
      jwt:
        type: string
    type: object
  models.LoginRequestModel:
    properties:
      password:
        type: string
      username:
        type: string
    required:
    - password
    - username
    type: object
...
  models.RegisterRequestModel:
    properties:
      email:
        type: string
      name:
        maxLength: 50
        minLength: 2
        type: string
      password:
        minLength: 6
        type: string
      phone:
        type: string
      surname:
        maxLength: 50
        minLength: 2
        type: string
      username:
        maxLength: 32
        minLength: 3
        type: string
    required:
    - email
    - name
    - password
    - phone
    - surname
    - username
    type: object
  models.SuccessResponse:
    properties:
      message:
        example: Successfully {action}
        type: string
    type: object
  models.UserInfoResponseModel:
    properties:
      created_at:
        type: string
      email:
        type: string
      name:
        type: string
      phone:
        type: string
      surname:
        type: string
      username:
        type: string
    type: object
  models.UserInfoUpdateRequestModel:
    properties:
      email:
        type: string
      name:
        maxLength: 50
        minLength: 2
        type: string
      phone:
        type: string
      surname:
        maxLength: 50
        minLength: 2
        type: string
    required:
    - email
    - name
    - phone
    - surname
    type: object
...
info:
  contact: {}
paths:
  /auth/login:
    post:
      consumes:
      - application/json
      description: Вход пользователя в систему и получение JWT токена
      parameters:
      - description: Данные для входа
        in: body
        name: request
        required: true
        schema:
          $ref: '#/definitions/models.LoginRequestModel'
      produces:
      - application/json
      responses:
        "200":
          description: Успешная аутентификация
          schema:
            $ref: '#/definitions/models.JWTModel'
        "400":
          description: Ошибка валидации
          schema:
            $ref: '#/definitions/models.ErrorResponse'
        "401":
          description: Неверные учетные данные
          schema:
            $ref: '#/definitions/models.ErrorResponse'
        "500":
          description: Внутренняя ошибка сервера
          schema:
            $ref: '#/definitions/models.ErrorResponse'
      summary: Аутентификация пользователя
      tags:
      - Auth
  /auth/register:
    post:
      consumes:
      - application/json
      description: Создание нового аккаунта пользователя
      parameters:
      - description: Данные для регистрации
        in: body
        name: request
        required: true
        schema:
          $ref: '#/definitions/models.RegisterRequestModel'
      produces:
      - application/json
      responses:
        "201":
          description: Пользователь успешно зарегистрирован
          schema:
            $ref: '#/definitions/models.SuccessResponse'
        "400":
          description: Ошибка валидации
          schema:
            $ref: '#/definitions/models.ErrorResponse'
        "409":
          description: Пользователь уже существует
          schema:
            $ref: '#/definitions/models.ErrorResponse'
        "500":
          description: Внутренняя ошибка сервера
          schema:
            $ref: '#/definitions/models.ErrorResponse'
      summary: Регистрация нового пользователя
      tags:
      - Auth
...
  /user/info:
    delete:
      description: Удаляет основную информацию о пользователе
      produces:
      - application/json
      responses:
        "200":
          description: Успешное удаление
          schema:
            type: string
        "400":
          description: Неверный формат ID
          schema:
            $ref: '#/definitions/models.ErrorResponse'
        "401":
          description: Отсутствует авторизация
          schema:
            $ref: '#/definitions/models.ErrorResponse'
        "500":
          description: Внутренняя ошибка сервера
          schema:
            $ref: '#/definitions/models.ErrorResponse'
      security:
      - BearerAuth: []
      summary: Удаление информации пользователя
      tags:
      - User Info
    get:
      description: Получает основную информацию о пользователе
      produces:
      - application/json
      responses:
        "200":
          description: Информация о пользователе
          schema:
            $ref: '#/definitions/models.UserInfoResponseModel'
        "400":
          description: Неверный формат ID
          schema:
            $ref: '#/definitions/models.ErrorResponse'
        "401":
          description: Отсутствует авторизация
          schema:
            $ref: '#/definitions/models.ErrorResponse'
        "500":
          description: Внутренняя ошибка сервера
          schema:
            $ref: '#/definitions/models.ErrorResponse'
      security:
      - BearerAuth: []
      summary: Получение информации пользователя
      tags:
      - User Info
    patch:
      consumes:
      - application/json
      description: Обновляет основную информацию о пользователе
      parameters:
      - description: Данные для обновления
        in: body
        name: request
        required: true
        schema:
          $ref: '#/definitions/models.UserInfoUpdateRequestModel'
      produces:
      - application/json
      responses:
        "200":
          description: Успешное обновление
          schema:
            type: string
        "400":
          description: Ошибка валидации или неверный формат ID
          schema:
            $ref: '#/definitions/models.ErrorResponse'
        "401":
          description: Отсутствует авторизация
          schema:
            $ref: '#/definitions/models.ErrorResponse'
        "500":
          description: Внутренняя ошибка сервера
          schema:
            $ref: '#/definitions/models.ErrorResponse'
      security:
      - BearerAuth: []
      summary: Обновление информации пользователя
      tags:
      - User Info
...
securityDefinitions:
  BearerAuth:
    description: Type "Bearer" followed by a space and JWT token.
    in: header
    name: Authorization
    type: apiKey
swagger: "2.0"
...